tests: consolidate functions
Changelog: none
